requirements.txt
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
confection==0.0.4
contourpy==1.0.6
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.5
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
fonttools==4.38.0
fqdn==1.5.1
fst-pso==1.8.1
funcy==2.0
FuzzyTM==2.0.5
gensim==4.3.1
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.19.4
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.2
jupyter_server==2.0.6
jupyter_server_terminals==0.4.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
langcodes==3.3.0
langdetect==1.0.9
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
miniful==0.0.6
mistune==2.0.4
murmurhash==1.0.9
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.1
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.2
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.3
packaging==22.0
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
preshed==3.0.8
prometheus-client==0.15.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.5
pyFUME==0.2.25
Pygments==2.14.0
pyLDAvis==3.4.1
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.4
pytz==2022.7
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scikit-learn==1.2.1
scipy==1.10.0
seaborn==0.12.2
Send2Trash==1.8.0
simpful==2.9.0
six==1.16.0
smart-open==6.3.0
sniffio==1.3.0
soupsieve==2.3.2.post1
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
terminado==0.17.1
thinc==8.1.7
threadpoolctl==3.1.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
typer==0.7.0
typing_extensions==4.5.0
tzdata==2023.3
uri-template==1.2.0
urllib3==1.26.13
wasabi==1.1.1
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.5
zipp==3.11.0
